requirements.txt
fastapi==0.115.0
uvicorn==0.31.1
pydantic==2.9.2
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
PyJWT==2.9.0
bcrypt==4.2.0